Use operator-required INT8 static zero points

+++ 07_vision_prefill_optimization/benchmark_w8a8_vision_linears.py
@@ -148,7 +148,7 @@ def benchmark_one(
     weight_int8_kn = pack_weight(weight_int8_nk, device=device, layout=weight_layout)
     weight_scale = weight_scale.to(device=device, dtype=torch.float32)
     input_scale = input_scale_cpu.repeat(spec.in_features).to(device=device, dtype=torch.float32)
-    input_zero_point = torch.zeros(spec.in_features, device=device, dtype=torch.float32)
+    input_zero_point = torch.zeros(spec.in_features, device=device, dtype=torch.int8)
     dequant_scale = torch_npu.npu_trans_quant_param(dequant_scale_cpu.to(device=device, dtype=torch.float32))
     quant_bias = quant_bias_cpu.to(device=device)
